requirements.txt
onnxruntime==1.10.0
Pillow==8.4.0
Shapely==1.8.1.post1
opencv-python-headless==4.5.5.64
rx==3.2.0
websocket-client==1.3.1
tornado==6.1
rel==0.4.7